Moved transformer et al. to grizzly_extras
