+++ grizzly/task.py
@@ -4,10 +4,12 @@ from typing import List, Optional, Callable, Any, Type
 from jinja2.environment import Template
 from gevent import sleep as gsleep
 
-from .types import ResponseContentType, HandlerType, RequestMethod
+from .types import HandlerType, RequestMethod
 from .context import GrizzlyContext, GrizzlyTask, GrizzlyTasksBase
-from .transformer import Transformer, transformer
-from .exceptions import TransformerError
+
+from grizzly_extras.transformer import Transformer, transformer
+from grizzly_extras.types import ResponseContentType
+from grizzly_extras.exceptions import TransformerError
 
 __all__ = [
     'RequestTaskHandlers',
