+++ grizzly/testdata/variables/messagequeue.py
@@ -60,11 +60,11 @@ import zmq
 
 from gevent import sleep as gsleep
 from grizzly_extras.async_message import AsyncMessageContext, AsyncMessageRequest, AsyncMessageResponse
+from grizzly_extras.transformer import transformer
+from grizzly_extras.exceptions import TransformerError
 
 from ...types import bool_typed, str_response_content_type, AtomicVariable
 from ...context import GrizzlyContext
-from ...transformer import transformer
-from ...exceptions import TransformerError
 from ..utils import resolve_variable
 
 try:
