+++ grizzly/users/meta/response_handler.py
@@ -4,11 +4,12 @@ from locust.clients import ResponseContextManager
 
 from ...task import RequestTask
 from ...types import HandlerContextType, ResponseContentType
-from ...transformer import transformer
-from ...exceptions import TransformerError, ResponseHandlerError
+from ...exceptions import TransformerLocustError, ResponseHandlerError
 from .context_variables import ContextVariables
 from .response_event import ResponseEvent
 
+from grizzly_extras.transformer import transformer
+from grizzly_extras.types import ResponseContentType
 
 class ResponseHandler(ResponseEvent):
     abstract = True
@@ -61,8 +62,8 @@ class ResponseHandler(ResponseEvent):
                             break
 
                 if response_content_type is ResponseContentType.GUESS:
-                    raise TransformerError(f'failed to transform: {response_payload}')
-            except TransformerError as e:
+                    raise TransformerLocustError(f'failed to transform: {response_payload}')
+            except TransformerLocustError as e:
                 if response_context is not None:
                     response_context.failure(e.message)
                     return
