+++ grizzly_extras/exceptions.py
@@ -1,0 +1,9 @@
+from typing import Optional
+
+
+class TransformerError(Exception):
+    message: Optional[str] = None
+
+    def __init__(self, message: Optional[str] = None) -> None:
+        self.message = message
+
